requirements.txt
anyio>=4.0.0
httpx>=0.25.0
mcp>=1.0.0
python-dotenv>=1.0.0
agent-framework
agent-framework-azure
azure-identity
